server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Physics RAG study assistant backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "langchain>=0.3.25",
    "langchain-ollama>=0.3.0",
    "langchain-openai>=0.3.16",
    "langchain-text-splitters>=0.3.0",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.11.0",
    "pypdf>=5.4.0",
    "bcrypt>=4.0.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.41",
    "sentence-transformers>=5.4.1",
    "langchain-google-genai>=4.2.2",
    "marker-pdf>=1.10.0",
    "llama-parse>=0.6.94",
    "llama-cloud-services>=0.6.94",
    "langchain-huggingface>=1.2.2",
    "langchain-community>=0.4.1",
]

[dependency-groups]
dev = [
    "fpdf2>=2.8.0",
    "pytest>=8.0.0",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
